+++ apps/commerce-api/src/main/java/com/loopers/application/MemberModifyService.java
@@ -5,7 +5,7 @@ import org.springframework.validation.annotation.Validated;
 
 import com.loopers.application.provided.MemberRegister;
 import com.loopers.application.required.MemberRepository;
-import com.loopers.domain.DuplicateMemberIdException;
+import com.loopers.domain.member.DuplicateMemberIdException;
 import com.loopers.domain.member.Member;
 import com.loopers.domain.member.MemberRegisterRequest;
