+++ apps/commerce-api/src/main/java/com/loopers/domain/member/DuplicateMemberIdException.java
@@ -1,4 +1,4 @@
-package com.loopers.domain;
+package com.loopers.domain.member;
 
 public class DuplicateMemberIdException extends RuntimeException {
     public DuplicateMemberIdException(String message) {}
