+++ apps/commerce-api/src/test/java/com/loopers/application/provided/MemberRegisterIntegrationTest.java
@@ -14,7 +14,7 @@ import org.springframework.test.context.bean.override.mockito.MockitoSpyBean;
 
 import com.loopers.application.MemberModifyService;
 import com.loopers.application.required.MemberRepository;
-import com.loopers.domain.DuplicateMemberIdException;
+import com.loopers.domain.member.DuplicateMemberIdException;
 import com.loopers.domain.member.Gender;
 import com.loopers.domain.member.Member;
 import com.loopers.domain.member.MemberFixture;
